pyproject.toml
[project]
name = "house-evaluator"
version = "0.1.0"
description = "Agentic home listing aesthetic evaluator"
readme = "CLAUDE.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.5",
    "openai>=1.0.0",
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
]

[project.scripts]
house = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
